requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-knox==5.0.1
djangorestframework==3.15.2
gTTS==2.5.1
gunicorn==22.0.0
idna==3.6
numpy==2.0.0
packaging==24.1
pillow==10.3.0
requests==2.31.0
setuptools==71.1.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
whitenoise==6.7.0
